Add support for automatic zero padding of inputs
This change implements automatic zero padding and modifies
two sampling functions such that the whole volume can be sampled.
Without the changes in this commit, the weighted and uniform sampling
were restricted such that the whole cropped input volume would fit within
the input volume. This is a breaking change, as it impacts the behavior
of the training.